introduce-request-validator(move): Hand inbound-request validation over to RequestValidator
Pure physical move per MECH_COMMIT_SPLIT. Cut the @staticmethod _validate_*
methods from TokenizerManager; paste into RequestValidator (drop
@staticmethod, replace ``self: "RequestValidator"`` -> plain ``self``).

Privacy flip per design (scope-induced rename — private TM helper becomes
public API of the new class):
  _validate_one_request -> validate_one
  _validate_input_ids_in_vocab -> validate_input_ids_in_vocab
  _validate_batch_tokenization_constraints -> validate_batch_tokenization_constraints
  _validate_mm_limits / _validate_for_matryoshka_dim stay private.

Caller prefix replacement:
  TokenizerManager._validate_one_request(self.request_validator, ...)
    -> self.request_validator.validate_one(obj=..., input_ids=...)
  TokenizerManager._validate_mm_limits(self.request_validator, ...)
    -> self.request_validator._validate_mm_limits(...)
  TokenizerManager._validate_batch_tokenization_constraints(self.request_validator, ...)
    -> self.request_validator.validate_batch_tokenization_constraints(batch_size=..., obj=...)
Internal call inside validate_one collapses
  TokenizerManager._validate_for_matryoshka_dim(self, obj)
    -> self._validate_for_matryoshka_dim(obj)

Refactor chain ID: introduce-request-validator-move
PR-Title: Extract RequestValidator from TokenizerManager
