+++ python/sglang/srt/managers/tokenizer_manager.py
@@ -570,7 +570,7 @@ class TokenizerManager(TokenizerControlMixin):
             if obj.audio_data is not None and not isinstance(obj.audio_data, list):
                 obj.audio_data = [obj.audio_data]
             if contains_mm_input:
-                TokenizerManager._validate_mm_limits(self.request_validator, obj)
+                self.request_validator._validate_mm_limits(obj)
 
             mm_inputs = None
 
@@ -662,161 +662,10 @@ class TokenizerManager(TokenizerControlMixin):
         else:
             mm_inputs = None
 
-        TokenizerManager._validate_one_request(self.request_validator, obj, input_ids)
+        self.request_validator.validate_one(obj=obj, input_ids=input_ids)
         return self._create_tokenized_object(
             obj, input_text, input_ids, input_embeds, mm_inputs, token_type_ids
         )
-
-    @staticmethod
-    def _validate_one_request(
-        self: "RequestValidator",
-        obj: Union[GenerateReqInput, EmbeddingReqInput],
-        input_ids: List[int],
-    ) -> None:
-        """Validates that the input token count and the requested token count doesn't exceed the model's context length."""
-        # FIXME: unify the length validation logic with the one in the scheduler.
-        _max_req_len = self.config.context_len
-        input_token_num = len(input_ids) if input_ids is not None else 0
-        input_token_num += self.config.num_reserved_tokens
-
-        # Validate input length
-        if input_token_num >= self.config.context_len:
-            if self.config.allow_auto_truncate:
-                logger.warning(
-                    f"The input ({input_token_num} tokens) is longer than the "
-                    f"model's context length ({self.config.context_len} tokens). "
-                    "Truncating the input."
-                )
-                del input_ids[_max_req_len:]
-                input_token_num = len(input_ids)
-            else:
-                raise ValueError(
-                    f"The input ({input_token_num} tokens) is longer than the "
-                    f"model's context length ({self.config.context_len} tokens)."
-                )
-
-        # Validate total tokens (input + max_new_tokens)
-        max_new_tokens = obj.sampling_params.get("max_new_tokens")
-        if (
-            self.config.validate_total_tokens
-            and max_new_tokens is not None
-            and (max_new_tokens + input_token_num) > _max_req_len
-        ):
-            if self.config.allow_auto_truncate:
-                logger.warning(
-                    f"Requested token count ({input_token_num} input + {max_new_tokens} new) "
-                    f"exceeds the model's context length ({self.config.context_len} tokens). "
-                    "Truncating max_new_tokens."
-                )
-                obj.sampling_params["max_new_tokens"] = max(
-                    0, _max_req_len - input_token_num
-                )
-            else:
-                total_tokens = max_new_tokens + input_token_num
-                error_msg = (
-                    f"Requested token count exceeds the model's maximum context length "
-                    f"of {self.config.context_len} tokens. You requested a total of {total_tokens} "
-                    f"tokens: {input_token_num} tokens from the input messages and "
-                    f"{max_new_tokens} tokens for the completion. Please reduce the number "
-                    f"of tokens in the input messages or the completion to fit within the limit."
-                )
-                raise ValueError(error_msg)
-
-        # Validate embedding requests
-        if isinstance(obj, EmbeddingReqInput) and self.config.is_generation:
-            raise ValueError(
-                "This model does not appear to be an embedding model by default. "
-                "Please add `--is-embedding` when launching the server or try another model."
-            )
-
-        # Validate Matryoshka embeddings
-        if isinstance(obj, EmbeddingReqInput):
-            TokenizerManager._validate_for_matryoshka_dim(self, obj)
-
-        # Validate custom logit processor
-        if isinstance(obj, GenerateReqInput):
-            if obj.return_hidden_states and not self.config.enable_return_hidden_states:
-                raise ValueError(
-                    "The server is not configured to return the hidden states. "
-                    "Please set `--enable-return-hidden-states` to enable this feature."
-                )
-            if (
-                obj.custom_logit_processor
-                and not self.config.enable_custom_logit_processor
-            ):
-                raise ValueError(
-                    "The server is not configured to enable custom logit processor. "
-                    "Please set `--enable-custom-logit-processor` to enable this feature."
-                )
-
-    @staticmethod
-    def _validate_mm_limits(
-        self: "RequestValidator",
-        obj: Union[GenerateReqInput, EmbeddingReqInput],
-    ) -> None:
-        if not self.config.limit_mm_data_per_request:
-            return
-
-        for modality, limit in self.config.limit_mm_data_per_request.items():
-            data = getattr(obj, f"{modality}_data", None)
-            if data:
-                count = len(data) if isinstance(data, list) else 1
-                if count > limit:
-                    raise ValueError(
-                        f"{modality.capitalize()} count {count} exceeds limit {limit} per request."
-                    )
-
-    @staticmethod
-    def _validate_for_matryoshka_dim(
-        self: "RequestValidator", obj: EmbeddingReqInput
-    ) -> None:
-        """Validate the request for Matryoshka dim if it has the field set."""
-        if obj.dimensions is None:
-            return
-
-        if not self.config.is_matryoshka:
-            raise ValueError(
-                f"Model '{self.config.model_path}' does not support matryoshka representation, "
-                f"changing output dimensions will lead to poor results."
-            )
-
-        if obj.dimensions < 1:
-            raise ValueError("Requested dimensions must be greater than 0")
-
-        if (
-            self.config.matryoshka_dimensions
-            and obj.dimensions not in self.config.matryoshka_dimensions
-        ):
-            raise ValueError(
-                f"Model '{self.config.model_path}' only supports {self.config.matryoshka_dimensions} matryoshka dimensions, "
-                f"using other output dimensions will lead to poor results."
-            )
-
-        if obj.dimensions > self.config.hidden_size:
-            raise ValueError(
-                f"Provided dimensions are greater than max embedding dimension: {self.config.hidden_size}"
-            )
-
-    @staticmethod
-    def _validate_input_ids_in_vocab(
-        self: "RequestValidator",
-        input_ids: Union[List[int], List[List[int]]],
-        vocab_size: int,
-    ) -> None:
-        # Handle both single sequence and batch of sequences
-        if isinstance(input_ids[0], list):
-            # Batch of sequences
-            for seq in input_ids:
-                if any(id >= vocab_size for id in seq):
-                    raise ValueError(
-                        f"The input_ids {seq} contains values greater than the vocab size ({vocab_size})."
-                    )
-        else:
-            # Single sequence
-            if any(id >= vocab_size for id in input_ids):
-                raise ValueError(
-                    f"The input_ids {input_ids} contains values greater than the vocab size ({vocab_size})."
-                )
 
     def _create_tokenized_object(
         self,
@@ -957,8 +806,8 @@ class TokenizerManager(TokenizerControlMixin):
             # All requests already have input_ids, no need to tokenize
             return [await self._tokenize_one_request(obj[i]) for i in range(batch_size)]
 
-        TokenizerManager._validate_batch_tokenization_constraints(
-            self.request_validator, batch_size, obj
+        self.request_validator.validate_batch_tokenization_constraints(
+            batch_size=batch_size, obj=obj
         )
 
         # Collect requests and texts
@@ -981,9 +830,7 @@ class TokenizerManager(TokenizerControlMixin):
         # Process all requests
         tokenized_objs = []
         for i, req in enumerate(requests):
-            TokenizerManager._validate_one_request(
-                self.request_validator, obj[i], input_ids_list[i]
-            )
+            self.request_validator.validate_one(obj=obj[i], input_ids=input_ids_list[i])
             token_type_ids = (
                 token_type_ids_list[i] if token_type_ids_list is not None else None
             )
@@ -994,27 +841,6 @@ class TokenizerManager(TokenizerControlMixin):
             )
         logger.debug(f"Completed batch processing for {batch_size} requests")
         return tokenized_objs
-
-    @staticmethod
-    def _validate_batch_tokenization_constraints(
-        self: "RequestValidator",
-        batch_size: int,
-        obj: Union[GenerateReqInput, EmbeddingReqInput],
-    ) -> None:
-        """Validate constraints for batch tokenization processing."""
-        for i in range(batch_size):
-            if self.config.is_generation and obj[i].contains_mm_input():
-                raise ValueError(
-                    "For multimodal input processing do not set `enable_tokenizer_batch_encode`."
-                )
-            if obj[i].input_ids is not None:
-                raise ValueError(
-                    "Batch tokenization is not needed for pre-tokenized input_ids. Do not set `enable_tokenizer_batch_encode`."
-                )
-            if obj[i].input_embeds is not None:
-                raise ValueError(
-                    "Batch tokenization is not needed for input_embeds. Do not set `enable_tokenizer_batch_encode`."
-                )
 
     def _batch_has_text(
         self, batch_size: int, obj: Union[GenerateReqInput, EmbeddingReqInput]
